@@ -19,14 +19,11 @@ class Ability
         $user = auth()->user();
 
         if (!($user->roles->isEmpty())
-            && in_array($permission, $user->permissions())
-            && 'trainer-provider' == $user->roles->toArray()[0]['code']
+            && ( (in_array($permission, $user->permissions()) && 'trainer-provider'  == $user->roles->toArray()[0]['code'])
+                || 'student'  == $user->roles->toArray()[0]['code'])
+
         ) {
             return $next($request);
-        }
-
-        if (!is_null($user)) {
-            return redirect()->route('blank-dashboard');
         }
 
         return redirect()->route('login');
