@@ -3,6 +3,7 @@
 namespace App\Repositories;
 
 use App\Models\Classes;
+use Carbon\Carbon;
 
 class ClassesRepository extends Repository
 {
@@ -10,6 +11,7 @@ class ClassesRepository extends Repository
 
     /**
      * ClassGroupRepository constructor.
+     *
      * @param Classes $classGroup
      */
     public function __construct(Classes $classGroup)
@@ -22,11 +24,18 @@ class ClassesRepository extends Repository
         $model = $this->model;
 
         if (auth()->user()->isTrainerProvider()) {
-            $model = $model->whereHas('mainClass', function($query) {
-                return $query->where('trainer_provider_id', auth()->user()->trainerProvider->getId());
-            });
+            $model = $model->whereHas(
+                'mainClass',
+                function ($query) {
+                    return $query->where('trainer_provider_id', auth()->user()->trainerProvider->getId());
+                }
+            );
         }
 
         return $model->orderBy($column, $order)->get();
     }
+
 }
+
+
+
