@@ -1,5 +1,6 @@
 <?php
 
+use Database\Seeders\SqlFileSeeder;
 use Illuminate\Database\Seeder;
 
 class DatabaseSeeder extends Seeder
@@ -11,6 +12,8 @@ class DatabaseSeeder extends Seeder
      */
     public function run()
     {
-        // $this->call(UserSeeder::class);
+        //$path = public_path('academia_2020.sql');
+        //$sql = file_get_contents($path);
+        //DB::unprepared($sql);
     }
 }
