@@ -1,6 +1,3 @@
-<div class="pull-right">
-    Made by <strong><a href="https://www.coredesign.ro/" target="_blank">CORE Design</a></strong>
-</div>
 <div>
-    <strong>Copyright</strong> Academia Testarii &copy; 2019 - 2020
+  <strong>Copyright</strong> Academia Testarii &copy; 2019 - 2020
 </div>
